odoo techms turned into blueprint
